requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
langgraph==0.0.66
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
gtts==2.4.0
pydub==0.25.1
python-dotenv==1.0.0
redis==5.0.1
websockets==12.0
aiofiles==23.2.1
httpx==0.25.2